Update utils to not be so fragile.
The old way that utils was looking to split the header was going off the first
item from split lines. Which isnt always the case. The new way will look for
the form seperator begining with a --- and use that as the new seperator.

After this it will remove and junk space from the lines and split the first new
line character it finds. Then create the header and content.