Created the package asynchronously.
The package creation can take a long time, specially for old
packages with many releases (e.g.: Celery). This causes the
simple_detail view to take too long to answer which also causes pip
to timeout and retry the connection, starting another slow process.

With this commit, the package is created asynchronously in the
Celery worker. I have also added an 'lock' in order to stop the
task from being queued more than once. This is required because
a `pip install` call does actually two requests on the simple_detail
view.